Change PostRequest API to support raw messages and JSON messages
In some circumstances, it's possible for POST requests not to be JSON. We want to be able to send
requests that contain raw bytes, which is what this API change does.

Also, actions that do not have parameters now do not take an empty struct. Instead they simply do
not take a "params" parameter