No more pointers to maps and slices and no constructors
There are new constructors that explicitly set the default values for structs, otherwise new struct
creation should happen simply through normal struct instantiation. Also, default values are not set
during serialization, they are only included if explicitly set in the original struct.

Pointers to maps and slices are just difficult to work with, so they are not used