Fix restli serialization bug for records
If multiple record fields are optional, the old solution was likely to add too many commas. To fix
this, some state is kept to indicate whether or not any field was written to the buffer and
therefore if a comma needs to be added.

Also add a cosmetic fix to the union field serializer to use a switch statement instead of a goto