Use a strings.Reader when creating the request
When using readers other than strings.Reader, bytes.Buffer or bytes.Reader, the http.Client will use
a chunked encoding because it technically does not know the size of the request upfront. The
io.ReadCloser returned by jwriter is _not_ one of those readers so it uses chunked encoding, which
makes it difficult to debug raw requests because they contain some garbage bytes to indicate chunk
boundaries. Using the ReadCloser from jwriter will read from the buffer pool directly instead of
copying the bytes out into a string. This can cause a lot of garbage in high-throughput services but
is hopefully negligible in most cases. This feature can be revived opt-in in the future if needed