Set the ContentLength field in the request to remove the chunked encoding
The chunked encoding is problematic because it makes debugging raw requests a little bit more
involved. On the other hand, copying all the bytes out of the reader will create lots of garbage so
it's undesirable. Fortunately setting the ContentLength field actually causes the http.Client to
read exactly that many bytes from the Reader and no use chunked encoding, which is a win-win.

Also produce a report of the test coverage for the suite at the end of the integration test